[backend] chore : ID 식별자 LONG으로 변경
[backend] chore : ID 식별자 LONG으로 변경

+++ backend/yourwine/src/main/java/com/ssafy/yourwine/model/entity/Wine.java
@@ -22,7 +22,7 @@ public class Wine {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "wine_id")
-	private int wineId;
+	private Long wineId;
 	private String img;
 	private String kname;
 	private String ename;
